requirements.txt
Flask==0.12.2
gunicorn==19.7.1
raven==6.4.0
blinker==1.4
xlwt==1.3.0
requests==2.18.4
Shapely==1.6.3
python-dateutil==2.6.1
flask-cors==3.0.3
